+++ lib/pos_printer_manager.dart
@@ -1,13 +1,14 @@
 // Auto-generated exports
 
 export 'src/manager.dart';
+export 'src/models/label_printer_language.dart';
 export 'src/models/pos_printer.dart';
 export 'src/models/printer_config.dart';
 export 'src/plugins/andro_bar_printer/andro_bar_printer_plugin.dart';
 export 'src/plugins/category_for_printer.dart';
 export 'src/plugins/kitchen_printer/kitchen_printer_plugin.dart';
 export 'src/plugins/label_printer/label_printer_plugin.dart';
-export 'src/plugins/label_printer/zpl_demo.dart';
+export 'src/plugins/label_printer/label_builder.dart';
 export 'src/plugins/printer_connection_params_extension.dart';
 export 'src/plugins/printer_settings.dart';
 export 'src/plugins/printer_type.dart';